+++ flutter_readium/example/lib/pages/bookshelf.page.dart
@@ -202,6 +202,8 @@ class BookshelfPageState extends State<BookshelfPage> {
       return 'Audiobook';
     } else if (pub.conformsToReadiumPDF) {
       return 'PDF';
+    } else if (pub.conformsToReadiumDivina) {
+      return 'Comic';
     } else {
       return 'Unknown format';
     }