+++ flutter_readium/example/lib/utils/publication_utils.dart
@@ -39,6 +39,8 @@ class PublicationUtils {
       '.webpub',
       '.epub',
       '.audiobook',
+      '.cbz',
+      '.divina',
       '.pdf',
       '.zip',
       '.json',